+++ src/main/java/com/exam/qna/entity/Question.java
@@ -8,7 +8,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
@@ -28,13 +27,6 @@ public class Question {
     private String content;
     private LocalDateTime createDate;
 
-
-    // ALL : 부모엔티티에 관련된 데이터가 저장될때 자식 엔티티도 저장을 할까 ?
-    @OneToMany(mappedBy = "question", cascade = {CascadeType.REMOVE, CascadeType.ALL})
-    private List<Answer> answerList = new ArrayList<>();
-
-    public void addAnswer(Answer answer){
-        answer.setQuestion(this);
-        getAnswerList().add(answer);
-    }
+    @OneToMany(mappedBy = "question", cascade = CascadeType.REMOVE)
+    private List<Answer> answerList;
 }
