--- src/main/java/com/exam/qna/repository/RepositoryUtil.java
@@ -1,17 +1,0 @@
-package com.exam.qna.repository;
-
-import jakarta.transaction.Transactional;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
-
-public interface RepositoryUtil {
-    @Transactional
-    @Modifying
-    @Query(value = "SET FOREIGN_KEY_CHECKS = 0", nativeQuery = true)
-    void disableForeignKeyChecks();
-
-    @Transactional
-    @Modifying
-    @Query(value = "SET FOREIGN_KEY_CHECKS = 1", nativeQuery = true)
-    void enableForeignKeyChecks();
-}
