--- src/main/resources/templates/question_list.html
@@ -1,16 +1,0 @@
-<h2>질문 리스트</h2>
-
-<table>
-    <thead>
-    <tr>
-        <th>제목</th>
-        <th>작성일시</th>
-    </tr>
-    </thead>
-    <tbody>
-    <tr th:each="question : ${questionList}">
-        <td th:text="${question.subject}"></td>
-        <td th:text="${question.createDate}"></td>
-    </tr>
-    </tbody>
-</table>
