+++ src/test/java/com/exam/qna/QuestionRepositoryTests.java
@@ -19,40 +19,38 @@ public class QuestionRepositoryTests {
     @Autowired
     private QuestionRepository questionRepository;
 
-
     private static int lastSampleDataId;
 
     @BeforeEach
     void beforeEach(){
-        QuestionRepositoryTests.clearData(questionRepository);
-        QuestionRepositoryTests.createSampleData(questionRepository);
+        clearData();
+        createSampleData();
     }
 
-    public static int createSampleData(QuestionRepository questionRepository) {
+    void createSampleData() {
         Question q1 = new Question();
         q1.setSubject("이게 무엇인가요?");
         q1.setContent("이것에 대해서 알고 싶습니다.");
         q1.setCreateDate(LocalDateTime.now());
-        questionRepository.save(q1);  // 첫번째 질문 저장
+        this.questionRepository.save(q1);  // 첫번째 질문 저장
 
         Question q2 = new Question();
         q2.setSubject("스프링부트 모델 질문입니다.");
         q2.setContent("id는 자동으로 생성되나요?");
         q2.setCreateDate(LocalDateTime.now());
-        questionRepository.save(q2);  // 두번째 질문 저장
+        this.questionRepository.save(q2);  // 두번째 질문 저장
 
-        return lastSampleDataId = q2.getId();
+        lastSampleDataId = q2.getId();
+
+
     }
 
-    // 전역으로 사용 할 수 있는 static
-    public static void clearData(QuestionRepository questionRepository) {
-        //questionRepository.disableForeignKeyChecks();
-        // delete -> ALTER TABLE question AUTO_INCREMENT = 1; 바꾼다.
-        questionRepository.deleteAll();
+    void clearData() {
+        questionRepository.disableForeignKeyChecks();
         // 초기화
         questionRepository.truncate();
         // 외래키를 다시 만든다
-       // questionRepository.enableForeignKeyChecks();
+        questionRepository.enableForeignKeyChecks();
     }
 
     // save
